refactor(server): extract typed models module
